backend/requirements.txt
# Web Framework & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database & ORM
sqlalchemy==2.0.25
psycopg2-binary==2.9.10
alembic==1.13.1

# Background Tasks & Caching
redis==5.0.1
celery==5.3.6

# Risk Engine & Analytics
scikit-learn==1.5.2
pandas==2.2.3
numpy<2
xgboost==2.0.3
shap==0.44.1
imbalanced-learn==0.12.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic-settings==2.1.0
pydantic[email]==2.6.0

# Kavya: Invoice Processing Pipeline
pdfplumber==0.11.0
Pillow==10.4.0
opencv-python-headless==4.10.0.84
pytesseract==0.3.10

# Hashing - pycryptodome has pre-built wheels for Python 3.13
pycryptodome==3.20.0

# Fuzzy matching - rapidfuzz has pre-built wheels, no compilation
rapidfuzz==3.6.1

# Environment variables
python-dotenv==1.0.0

# S3 storage (KYC uploads)
boto3
